+++ app/src/js/reducers/datepicker.js
@@ -5,15 +5,17 @@ import {
   DATEPICKER_DROPDOWN_FILTER,
   DATEPICKER_HOUR_FORMAT
 } from '../actions/types';
-
-const daysToMilliseconds = 1000 * 60 * 60 * 24;
+import { msPerDay, allDateRanges } from '../utils/datepicker';
 
 // Also becomes default props for Datepicker
-export const initialState = {
-  startDateTime: null,
-  endDateTime: null,
-  dateRange: {value: 'All', label: 'All'},
-  hourFormat: '12HR'
+export const initialState = () => {
+  const now = new Date(Date.now());
+  return {
+    startDateTime: new Date(now - msPerDay),
+    endDateTime: null,
+    dateRange: allDateRanges.find((a) => a.value === 'Recent'),
+    hourFormat: '12HR'
+  };
 };
 
 /**
@@ -30,19 +32,37 @@ const computeDateTimeDelta = (timeDeltaInDays) => {
   if (!isNaN(timeDeltaInDays)) {
     endDateTime = new Date(Date.now());
     startDateTime = new Date(
-      endDateTime - timeDeltaInDays * daysToMilliseconds
+      endDateTime - timeDeltaInDays * msPerDay
     );
   }
   return { startDateTime, endDateTime };
 };
 
-export default function reducer (state = initialState, action) {
+/**
+* Sets the state for recent data start time is 24 hours ago, end time is null
+*
+* @returns {Object} with startDateTime and dateRange set to "Recent"
+*/
+const recentData = () => {
+  const endDateTime = null;
+  const startDateTime = new Date(Date.now() - msPerDay);
+  return {startDateTime, endDateTime, dateRange: allDateRanges.find((a) => a.value === 'Recent')};
+};
+
+export default function reducer (state = initialState(), action) {
   state = { ...state };
   const { data } = action;
   switch (action.type) {
     case DATEPICKER_DROPDOWN_FILTER:
-      state = {...state, ...computeDateTimeDelta(data.dateRange.value), ...data};
-      break;
+      switch (data.dateRange.label) {
+        case 'Custom':
+        case 'All':
+          return {...state, ...data, ...{startDateTime: null, endDateTime: null}};
+        case 'Recent':
+          return {...state, ...data, ...recentData()};
+        default:
+          return {...state, ...computeDateTimeDelta(data.dateRange.value), ...data};
+      }
     case DATEPICKER_DATECHANGE:
       state = { ...state, ...data };
       break;